requirements.txt
Flask
Flask-Login
pymongo
cryptography
python-dotenv
werkzeug